REFACTOR: corsOprions into its own class
